Refactor Flask app to use single SQLAlchemy instance
