+++ app/__init__.py
@@ -2,18 +2,25 @@ import os
 from datetime import timedelta
 
 from flask import Flask
+from flask_migrate import Migrate
+from flask_sqlalchemy import SQLAlchemy
 
-from app.models import db, User
-
-try:
+try:  # flask-login is optional in some environments
     from flask_login import LoginManager
 except Exception:  # pragma: no cover - flask_login optional
     LoginManager = None  # type: ignore[assignment]
 
-login_manager = LoginManager() if "LoginManager" in globals() and LoginManager else None
+# ---------------------------------------------------------------------------
+# Extension instances (singletons shared across the app)
+# ---------------------------------------------------------------------------
+db: SQLAlchemy = SQLAlchemy()
+migrate: Migrate = Migrate()
+login_manager = LoginManager() if LoginManager else None
 
 
 def create_app(config_overrides: dict | None = None) -> Flask:
+    """Application factory used by the web app, CLI, and tests."""
+
     app = Flask(
         __name__,
         static_url_path="/static",
@@ -21,7 +28,9 @@ def create_app(config_overrides: dict | None = None) -> Flask:
         template_folder="templates",
     )
 
-    default_db = os.getenv("DATABASE_URL") or f"sqlite:///{os.path.join(app.root_path, 'shardbound.db')}"
+    os.makedirs(app.instance_path, exist_ok=True)
+    default_db = os.getenv("DATABASE_URL") or f"sqlite:///{os.path.join(app.instance_path, 'shardbound.db')}"
+
     app.config.update(
         SECRET_KEY=os.getenv("SECRET_KEY", "dev-change-me"),
         SQLALCHEMY_DATABASE_URI=default_db,
@@ -30,12 +39,19 @@ def create_app(config_overrides: dict | None = None) -> Flask:
         SESSION_COOKIE_HTTPONLY=True,
         SESSION_COOKIE_SAMESITE="Lax",
     )
+
     if config_overrides:
         app.config.update(config_overrides)
 
     db.init_app(app)
 
+    # Import models after the database has been initialized so Alembic can
+    # detect metadata from the same SQLAlchemy instance.
+    from . import models  # noqa: F401
+
     if login_manager:
+        from .models import User
+
         login_manager.init_app(app)
         login_manager.login_view = "auth.login_page"
         login_manager.session_protection = "strong"
@@ -49,16 +65,19 @@ def create_app(config_overrides: dict | None = None) -> Flask:
             except (ValueError, TypeError):
                 return None
 
+    migrate.init_app(app, db)
+
     from .main.routes import main_bp
     from .game.routes import game_bp
     from .auth.routes import auth_bp
+    from .routes.api_classes import bp as classes_bp
 
     app.register_blueprint(main_bp)
     app.register_blueprint(game_bp)
     app.register_blueprint(auth_bp)
-
-    from .routes.api_classes import bp as classes_bp
-
     app.register_blueprint(classes_bp)
 
     return app
+
+
+__all__ = ["create_app", "db", "migrate", "login_manager"]
