+++ app/models.py
@@ -3,15 +3,13 @@ from __future__ import annotations
 
 from datetime import datetime, date
 import bcrypt
-from flask_sqlalchemy import SQLAlchemy
+from app import db
 
 try:
     from flask_login import UserMixin
 except Exception:  # pragma: no cover - flask_login optional
     class UserMixin:  # type: ignore[dead code]
         pass
-
-db = SQLAlchemy()
 
 class User(UserMixin, db.Model):
     __tablename__ = "users"
